Move IDependabotConfig to task toot
